+++ dvd_track.h
@@ -8,7 +8,7 @@
  * @param vmg_ifo dvdread handler for primary IFO
  * @return IFO number
  */
-uint8_t dvd_track_ifo_number(const ifo_handle_t *vmg_ifo, const uint16_t track_number);
+uint16_t dvd_track_ifo_number(const ifo_handle_t *vmg_ifo, const uint16_t track_number);
 
 uint8_t dvd_track_ttn(const ifo_handle_t *vmg_ifo, const uint16_t track_number);
